Move image-conversion-only modules to dataset/_image_conversion
Splits webknossos/dataset/_utils/ into the general-purpose helpers that
stay (buffered_slice_reader/writer, tensorstore_helpers, tinywkw — all
used well beyond conversion, e.g. by MagView reads/writes, agglomerate
attachments, export_as_tiff) and the modules that only exist to back
Dataset.from_images/add_layer_from_images, moved to a new private package:

- image_conversion.py, chunked_images.py, ims_chunked_images.py,
  mrc_chunked_images.py
- pims_images.py, pims_czi_reader.py, pims_dm_readers.py, pims_tiff_reader.py
- segmentation_recognition.py, infer_bounding_box_existing_files.py
- vendor/dm3.py, vendor/dm4.py

The new package sits at the same nesting depth as _utils, so none of the
moved files' own relative imports needed changes — only external
references (dataset.py, a handful of tests, and the pyproject.toml ruff
exclude for vendor/, including a stale mirrored copy in
cluster_tools/pyproject.toml).
