+++ webknossos/webknossos/dataset/_image_conversion/chunked_images.py
