+++ webknossos/webknossos/dataset/_image_conversion/ims_chunked_images.py
