+++ webknossos/webknossos/dataset/_image_conversion/infer_bounding_box_existing_files.py
