+++ webknossos/webknossos/dataset/_image_conversion/mrc_chunked_images.py
