+++ webknossos/webknossos/dataset/_image_conversion/pims_czi_reader.py
