+++ webknossos/webknossos/dataset/_image_conversion/pims_dm_readers.py
