+++ webknossos/webknossos/dataset/_image_conversion/pims_images.py
