+++ webknossos/webknossos/dataset/_image_conversion/pims_tiff_reader.py
