+++ webknossos/webknossos/dataset/_image_conversion/segmentation_recognition.py
